SVG 버그 수정
https://github.com/openNAMU/openNAMU/issues/2008
