+++ app.py
@@ -2,8 +2,6 @@
 import os
 import re
 import logging
-import shutil
-import datetime
 
 from route.tool.func import *
 from route import *
@@ -201,69 +199,12 @@ with get_db_connect() as conn:
 
         server_set[i] = server_set_val
 
-    # Init-DB_care
-    if data_db_set['type'] == 'sqlite':
-        def back_up(back_time, back_up_where, back_up_count = 0):
-            try:
-                if back_up_count != 0:
-                    file_dir = os.path.split(back_up_where)[0]
-                    file_dir = '.' if file_dir == '' else file_dir
-                    
-                    file_name = os.path.split(back_up_where)[1]
-                    file_name = re.sub(r'\.db$', '_[0-9]{14}.db', file_name)
+        conn.commit()
 
-                    backup_file = [for_a for for_a in os.listdir(file_dir) if re.search('^' + file_name + '$', for_a)]
-                    backup_file = sorted(backup_file)
-                    
-                    if len(backup_file) >= back_up_count:
-                        remove_dir = os.path.join(file_dir, backup_file[0])
-                        os.remove(remove_dir)
-                        print('Back up : Remove (' + remove_dir + ')')
+auto_do_something(data_db_set)
 
-                now_time = datetime.datetime.now().strftime('%Y%m%d%H%M%S')
-                new_file_name = re.sub(r'\.db$', '_' + now_time + '.db', back_up_where)
-                shutil.copyfile(
-                    data_db_set['name'] + '.db', 
-                    new_file_name
-                )
-
-                print('Back up : OK (' + new_file_name + ')')
-            except Exception as e:
-                print('Back up : Error')
-                print(e)
-
-            threading.Timer(
-                60 * 60 * back_time, 
-                back_up,
-                [back_time, back_up_where, back_up_count]
-            ).start()
-
-        curs.execute(db_change('select data from other where name = "back_up"'))
-        back_time = curs.fetchall()
-        back_time = float(number_check(back_time[0][0], True)) if back_time and back_time[0][0] != '' else 0
-
-        curs.execute(db_change('select data from other where name = "backup_count"'))
-        back_up_count = curs.fetchall()
-        back_up_count = int(number_check(back_up_count[0][0])) if back_up_count and back_up_count[0][0] != '' else 0
-
-        if back_time != 0:
-            curs.execute(db_change('select data from other where name = "backup_where"'))
-            back_up_where = curs.fetchall()
-            back_up_where = back_up_where[0][0] if back_up_where and back_up_where[0][0] != '' else data_db_set['name'] + '.db'
-
-            print('Back up state : ' + str(back_time) + ' hours')
-            print('Back up directory : ' + back_up_where)
-            if back_up_count != 0:
-                print('Back up max number : ' + str(back_up_count))
-
-            back_up(back_time, back_up_where, back_up_count)
-        else:
-            print('Back up state : Turn off')
-
-    print('Now running... http://localhost:' + server_set['port'])
+print('Now running... http://localhost:' + server_set['port'])
     
-    conn.commit()
-
 # Init-custom
 if os.path.exists('custom.py'):
     from custom import custom_run
