+++ version.json
@@ -1,6 +1,6 @@
 {
     "beta" : {
-        "r_ver" : "v3.4.6-RC5-dev62",
+        "r_ver" : "v3.4.6-RC5-dev63",
         "c_ver" : "3500373",
         "s_ver" : "3500112"
     }
